Entity-Component system rewrite
It's basically done, but there's still some open questions/future work.

The ones that come immediately to mind:

- the save contains {}/nil array entries for each component in each
entity. The component list and _TYPE are overwritten on load anyways; we
should omit them from the save file. This will require either creating
and saving temporary objects or changing the repr() api to permit
filtering, I think -- repr_with() isn't powerful enough as is, since it
outputs the array part unconditionally even if it consists of nils.

- The API for calling and returning from component methods still needs
some thought. Right now the method is just called with self and has to
access its data as self.<component name>.foo. Do we want to pass in the
component as the first argument after self? And either way, is "first
method to return wins" a good way of handling returns?

- Attempting to access a nonexistent field returns a method searcher
(i.e. a function). Calling it will silently do nothing. Attempting to
access a nonexistent field should instead error. This requires scanning
the component array at __index time and returning the method only if we
find something -- maybe we can cache the results?