Fix issue where settings didn't work at all
This is because the new __index approach to the TreeNodes meant that
node:set(...) was setting the .value in the TreeNode proxy, not in the
original Setting object. For now a __newindex has been installed that
deals with this by updating the original if the key being set already
exists in the original, but I'm not 100% sure this is desireable.