Move all UI/rendering to the new Window system
It's slow as hell, partly because it renders the entire playfield every
"frame" even when in menus. This could be sped up by having some way for
children to disable rendering in the parent, but I'm honestly not sure
how much of this is stuff I can/should optimize and how much is
Konsole's fault.