Massive UI library overhaul
Windows are now created orphaned and attached later. Attaching a Window
causes :reposition() to be called on it, which calls :resize() to
calculate preferred width and height and then sets actual width and
height based on the 'position' field and the dimensions of the parent.
At the moment only position='fixed' and position='center' are supported.