Input handling rewrite complete.
Key input is now read at the top level (in the love.main provided by
lovecompat) and passed to love.keypressed. Key events are passed to the
topmost window and percolate down until something handles them by (a)
having an appropriate event handler that (b) returns true.

Tree now has some really ugly hacks in it and a few user-visible bugs,
like not hilighting the active node immediately after creation, but
those will be dealt with during the Tree rewrite so that it operates
using a List for the renderer, which is up next.