Reworking of Position component and position ownership
Entity positioning is now handled by the map, which maintains a table of
(entity => position) information in addition to (position => entities),
and is responsible for keeping it up to date. The Position component
still exists, but only on Player, for implementing move/moveTo; this may
change in the future.

Entities created with map ownership (which, right now, is everything)
must have a defined position; they are created using Map:createAt()
rather than Map:create().