Separate drawing region from clipping region in tty lib
Previously, the clipping region and the drawing region were always eqv,
and any window pushed needed to be a subset of the current drawing
region. Attempting to push a window which this didn't apply do, or
drawing outside the current clipping region, resulted in an error.

With this change, the following semantics apply:
- The initial drawing region == the initial clipping region == the
terminal dimensions
- A pushed window can be of any size and position, even if this places
it outside the current clipping region or even outside the terminal
- When you push a window, the new clipping region is computed by
intersecting the current clipping region with the new window; the
result may be empty (i.e. 0x0)
- Draws outside the clipping region are ignored.

This makes scrolling, in particular, much simpler to implement.