WIP: make all children of a Tree focusable
This lets you scroll through all the children of the tree and updates
the scrolling in a mostly correct way, but includes invisible (but
.visible) children like VLists and Expanders, so there are "dead spots"
where you're pressing up/down and the focus just vanishes, because
you've focused one of these organizational elements rather than just an
individual TextLine or EntityLine.

Probably the solution here is to have the initial tree walk annotate
leaf nodes with a "can_focus" property and build the focus list based on
that.