Properly generate focus list in Tree
Previously, every visible child was added to the focus list. This
resulted in "dead" focus points when things like VLists were focused.
With this change, only windows with the can_focus property will be added
to the list. This property is set by default on all TextLines and
EntityLines; other subwindows being manually inserted into a Tree should
set it explicitly.